docs/syshealth-finalrelease-doc/final_release_code/reciever_side/requirements.txt
awscli==1.18.69
boto3==1.17.36
digi-xbee==1.3.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gspread==3.7.0
json==2.0.9
PyDrive==1.3.1